Build for Mac in the Release configuration at -O3
This is an experiment to get a few test cycles in so that we can measure the
impact and evaluate it against the .app/.dmg size increase that it will cause.
The size increase relative to -O2 is 1900kB (6%) .app and 800kB (5%) .dmg.

BUG=16713
TEST=Watch perf
Review URL: http://codereview.chromium.org/173505

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7cc6c0ba59df63f9f084f8cdfc695fa7b4ae1823