Mac cloud print proxy work.
Turn on use_cups for the Mac, add in the required library, and tweak the code
to get it linking.  There is still more work to do; this is just a gyp-oriented
change to get things started.

BUG=none
TEST=All targets still build on all platforms.

Review URL: http://codereview.chromium.org/4167006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 71be19b75c1366de0387a78f4623df2f5bf21745