Adding a Wayland basic toolkit
This is essentially a OO wrapper over the Wayland library. It will be used to
add Wayland support for Chrome.

This was written with the intent of being as standalone as possible and it 
should not require any external Chrome dependencies.

BUG=None
TEST=None

R=evan@chromium.org

Review URL: http://codereview.chromium.org/7457023

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 236754afe8615dfffebc8f5f4b0257c4cebcf137