restore OutputFile for chrome.exe, do moving hardlink on dll instead
Hardlink from different dir for chrome.exe caused annoyance when running from
Visual Studio because $(TargetPath) the default in the .user file is not the
correct path. So, instead do the hardlink on the dll. To avoid the possible
expense on non-NTFS, and to enable a trybot in the future, do this hardlink
based on a gyp define.

BUG=92528, 92685
TEST=No link problems


Review URL: http://codereview.chromium.org/7670031

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5ab8f48e3edea9bfa6a26bdf8b7767f62e76b82a