Add gamepad hardware data fetcher, and higher level thread container
DataFetcher is the hardware and platform-specific data getter that talks to OS
level interfaces for each platform and fill in a structure provided to it.
Currently only Windows is here.

Provider is a manager for the background thread that the DataFetcher runs on,
and interacts with the rest of browser.

This is part of a larger patch, the remainder of which is at
http://codereview.chromium.org/8345027/

BUG=79050


Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=111189

Review URL: http://codereview.chromium.org/8568029

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 58e56144d4831f31dbd4ea534d538bc7ae75fe03