Fix lastchange.py when running on native win32 python
When running as runhooks, it's not run via cygwin like it was when run in the build, and isn't able to find git. I believe the test that checked for both cygwin and win32 before was incorrect (but was always run via cygwin before so didn't matter).

Additionally, when runhooks runs it, it's from above src/, so the cwd isn't in the repo. So, pass in the directory of the script itself so that git can find its repo.

BUG=112264

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=120089

Review URL: https://chromiumcodereview.appspot.com/9317037

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fc127acd4a2b24b728f78e3f7975ad54548b9209