Lots of tweaks to the install-chroot.sh and install-build-deps.sh script to make them more userfriendly.
In particular, we now recognize systems that have multiple filesystems
(e.g. /home on NFS) and offer to do the right thing.

We also have better support for new Ubuntu distributions (both as
host and as guest). This means, we can now test on "precise".

Added a lot of extra error handling to catch common problems and either
fix them or offer suggestions on how the user can fix them. For example,
we now detect if the user tries to re-install the same chroot environment
multiple times; and we then offer to delete or overwrite the old
installation.

We also detect if a chroot environment is still in active use, and then
refuse to damage it.

In order to help users, who accidentally left an old chroot enviroment
running, we have added a "clean up" option to the wrapper script.

We automatically invoke install-build-deps.sh from install-chroot.sh,
so users only need to run install-chroot.sh and then answer a couple
of questions.

BUG=none
TEST=run install-chroot.sh on a Ubuntu machine and install both lucid32 and precise32. Verify that afterwards, we can build Chrome inside the chroot.
Review URL: https://chromiumcodereview.appspot.com/9702047

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 757c2962c2124c88aa73661ff698f04b6355cf4d