The background is that we are to enable hardfp build for  chrome browser.
Currently, the '-mfloat-abi=soft' is hard coded in the gyp file, but what we need is "-mfloat-abi=hard", since it is hard-coded, we no way to change it.

So this CL removed this hard-coded value, replaced it with variable arm_float_abi, which has a default value of 'soft'.   Later in the ebuild file, we can easily override this value to be 'hard', thus make the build hardfp.

BUG=None
TEST=build manually in chromium


Review URL: http://codereview.chromium.org/10351018

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5252af7ca157bf61fdf3cd67997d6b62fe99400f