Extracted inner class doing process dispatch.
The inner class WebResourceService::UnpackerClient is dispatching a json parsing job
to a subprocess. In iOS launching a subprocess is not supported. In order to replace
this class on this particular platform by one doing a thread dispatch the class has
been renamed and refactored in its own file. In doing so an abstract superclass was
crated to make the API more obvious.

The iOS subclass already exists in the iOS tree, and will be brought up with the
rest of the iOS browser changes.


BUG=b/6753678
TEST=No user visible changes.


Review URL: https://chromiumcodereview.appspot.com/10690096

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7e2d6e1cdbc0f6c05c5db1262f9d2a8d2eec1edb