export ANDROID_SDK_HOME when creating AVD
ANDROID_SDK_HOME is set in build/android/envsetup.sh, so android will
try to search for AVDs in that folder when running tests.

If not set this variable in this script, the AVDs created with this script
are invisible for the tests.

BUG=
TEST=


Review URL: https://chromiumcodereview.appspot.com/10782024

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 91ce11732f0d6aec110309a95d68cecc06f64d57