Android WebView: tidy up source/gyp layout.
1) Rename WebViewMainDelegate to AwMainDelegate for consistency with
other classes.

2) Move gyp file to the top level and don't bother having separate gypi
files for the subcomponents for now; the file list is very short anyway
and is not going to be a significant source of contention between the
small number of webview developers at this time, and it's easier to see
what's going on if it's all in one place.

BUG=


Review URL: https://chromiumcodereview.appspot.com/10893025

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c1a47df6521776547b5a3a9966dba63aab9e2e6b