Add initial Chromium TestShell support for Android
Add a build target and tie it into android_all.  This will build and link
Chrome code for a test apk for Android.  When it runs it properly
loads the library but crashes.  Will fix the crashes in subsequent
patches, but can fork that work after this is in.

BUG=136786


Review URL: https://chromiumcodereview.appspot.com/10918113

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 22f3f53880e2b3ff658d36a1fa6e214bb78ed3b0