Revert 156308 - Revert 156233 - Make apks be built in a consistent way
This broke non-incremental Android builds.

In the original CL there was a discrepancy in the way that Gyp handled
paths and the way that Ant handled paths. Changed it so that
resource_dir is specified relative to java_in_dir so that Gyp and Ant
treat it the same. For http://crbug.com/145991 I will have to make
resource_dir be in a consistent location for all apk targets and at that
point it can be removed from the gyp files.

Original:https://chromiumcodereview.appspot.com/10914130

BUG=


Review URL: https://chromiumcodereview.appspot.com/10913225

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e44c301ce6cb990c5ee4d9f11e34867d373d2b9