Always reference newly created script using full path.
When running build/install-chroot.sh to setup a Lucid64 environment for
testing on a Ubuntu Precise box the script fails with
  sudo: lucid64: command not found

This is because in Ubuntu Precise root does not have /usr/local/bin in
the path. Most references use the full path, this changes the remaining
to do so too.

BUG=155451


Review URL: https://chromiumcodereview.appspot.com/11187019

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d818503bb762833c52299f30d970295f1d72f262