Selective build clobbering feature (landmines.py and android build scripts).
Adds the ability for devs/troopers/etc. to set 'landmines' in the tree so that
the build will selectively clobber when a builder moves over a revision with such
a change.

This cl has an basis landmines.py, and hooks the clobber mechanism to the android
build scripts. 

The relevant cl which implements this for
compile.py is here: https://chromiumcodereview.appspot.com/11234013/

I'm planning to also implement an informational invocation for gclient to let devs know
about any potential landmines so they can decide if they need to clobber.

R=cmp,maruel@chromium.org
BUG=121897

Review URL: https://chromiumcodereview.appspot.com/11175016

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: da2f7cce4f1bec439e1a22f9b2f700991edace7b